requirements.txt
Coffin==2.0.1
contextlib2==0.5.4
defusedxml==0.4.1
Django==1.10.3
django-allauth==0.28.0
django-appconf==1.0.2
django-avatar==3.1.0
django-bower==5.2.0
django-faker==0.2
djangorestframework==3.5.3
fake-factory==9999.9.9
Faker==0.7.7
Jinja2==2.8
MarkupSafe==0.23
oauthlib==2.0.0
optional-django==0.1.0
Pillow==3.4.2
psycopg2==2.6.2
python-dateutil==2.6.0
python-openid==2.2.5
python3-openid==3.0.10
raven==5.32.0
requests==2.12.1
requests-oauthlib==0.7.0
six==1.10.0
